refactor(web): organize docs components
