+++ apps/web/src/docs/components/diagrams/CompactionMachineDiagram.tsx
