+++ apps/web/src/docs/components/diagrams/ComponentDiagram.tsx
