+++ apps/web/src/docs/components/diagrams/HarnessDiagram.tsx
