+++ apps/web/src/docs/components/diagrams/InterfaceComparisonDiagram.tsx
