+++ apps/web/src/docs/components/diagrams/MethodDiagram.tsx
