+++ apps/web/src/docs/components/diagrams/PrimitiveDiagram.tsx
