+++ apps/web/src/docs/components/diagrams/TransitionLoop.tsx
